honor root_path for location of chef installer script
If /tmp is mounted with noexec, then the `chmod` and `sudo the_script`
that are attempted later fail. Instead of hard-coding the location for
the script, write and run it from the provisioner's configured
root_path.

This fixes #1364 for the code path when `product_name` is present in the
Chef provisioner config.

Signed-off-by: Robb Kidd <robb@thekidds.org>